Completed TypeScript Conversion
