+++ types/types.ts
@@ -6,3 +6,13 @@ export interface MainPageBlockProps {
 	mainText: string;
 	buttonText: string | undefined;
 }
+
+export interface MainPageBlock2Props {
+	containerID: string;
+	imageSource: string;
+	imageAlt: string;
+	header: string;
+	mainText: string;
+	buttonText: string;
+	buttonHREF: string;
+}
